+++ release_notes.md
@@ -1,5 +1,36 @@
-﻿![Microchip logo](https://raw.githubusercontent.com/wiki/Microchip-MPLAB-Harmony/Microchip-MPLAB-Harmony.github.io/images/microchip_logo.png)
+---
+title: Release notes
+nav_order: 99
+---
+
+![Microchip logo](https://raw.githubusercontent.com/wiki/Microchip-MPLAB-Harmony/Microchip-MPLAB-Harmony.github.io/images/microchip_logo.png)
 ![Harmony logo small](https://raw.githubusercontent.com/wiki/Microchip-MPLAB-Harmony/Microchip-MPLAB-Harmony.github.io/images/microchip_mplab_harmony_logo_small.png)
 
 # Microchip MPLAB® Harmony 3 Release Notes
+## Microchip MPLAB® Harmony 3 Motor Control Application Examples for SAMRH71 family v3.0.0
 
+
+### Applications
+
+Applications migrated from motor_control repository to this application repository for SAMRH71 family. 
+
+| Development Board | Number of Applications | 
+| --- | --- | 
+|[SAMRH71F20-EK Evaluation Kit](https://www.microchip.com/DevelopmentTools/ProductDetails/PartNO/SAMRH71F20-EK) <br> [LX7720 Daughter Board](https://www.microsemi.com/product-directory/space-system-managers/3708-position-motor-controller-ic#resources) | 1 |
+|||
+
+
+### Required MPLAB Harmony v3 Modules
+* csp v3.8.3
+* x2c v1.1.3
+* motor_control v3.6.0
+* dev_packs v3.8.0
+* mhc v3.6.5
+
+### Development Tools
+
+* [MPLAB X IDE v5.45](https://www.microchip.com/mplab/mplab-x-ide)
+* [MPLAB XC32 C/C++ Compiler v2.50](https://www.microchip.com/mplab/compilers)
+* MPLAB X IDE plug-ins:
+  * MPLAB Harmony Configurator (MHC) v3.6.2
+  * X2CScope v1.3.0.